🚸 allow forward declaration of DD Package
