+++ include/dd/Operations.hpp
@@ -10,12 +10,6 @@
 
 #include <variant>
 
-namespace qc {
-using VectorDD = dd::vEdge;
-using MatrixDD = dd::mEdge;
-using DensityMatrixDD = dd::dEdge;
-} // namespace qc
-
 namespace dd {
 // single-target Operations
 template <class Config>
